Adds army rule generation and checking UI
Adds a new admin component to generate and check army rules, and upload them to Firebase for faster access.

The component provides a UI to trigger the generation of army rules for all factions and warband variants and also allows checking the status of pre-compiled rules.

The debug tools in the equipment selection dialog are now only visible to admins.